[FIX]: Move file to ViewModels
