--- GalleryCart/Areas/Artist/Models/IndexModel.cs
@@ -1,10 +1,0 @@
-﻿using GalleryCart.Models.Models;
-
-namespace GalleryCart.Areas.Artist.Models
-{
-    public class IndexModel
-    {
-        public User? CurrentUser { get; set; }
-        public required IQueryable<Post> Posts { get; set; }
-    }
-}
