--- GalleryCart/Areas/Customer/Models/CommissionManagementViewModel.cs
@@ -1,9 +1,0 @@
-﻿using GalleryCart.Models.Models;
-
-namespace GalleryCart.Areas.Customer.Models
-{
-    public class CommissionManagementViewModel
-    {
-        public Dictionary<User, List<Commission>> CommissionKvp{ get; set;} = new Dictionary<User, List<Commission>>();
-    }
-}
