+++ GalleryCart/Areas/Customer/Views/Commission/CommissionManagement.cshtml
@@ -1,6 +1,5 @@
-﻿@using GalleryCart.Areas.Customer.Models
-@using GalleryCart.Utilities.Utils
-@model CommissionManagementViewModel
+﻿@using GalleryCart.Utilities.Utils;
+@model GalleryCart.Models.ViewModels.CommissionManagementViewModel
 
 @{
     ViewData["Title"] = "All Your Commissions";
