+++ GalleryCart/Areas/Customer/Views/Commission/Create.cshtml
@@ -2,7 +2,7 @@
     ViewData["Title"] = "Commission Form";
     Layout = "~/Views/Shared/_Layout.cshtml";
 }
-@model GalleryCart.Areas.Customer.Models.CreateCommssionViewModel
+@model GalleryCart.Models.ViewModels.CreateCommssionViewModel
 
 <div class="min-h-screen bg-gray-800 text-white font-sans p-5">
     <div class="max-w-2xl mx-auto bg-gray-700 rounded-lg p-8 shadow-2xl">
