+++ GalleryCart/Views/Shared/CustomComponent/_Navbar.cshtml
@@ -189,7 +189,7 @@
                             <div class="flex space-x-4">
                                 <!-- Guest Navigation -->
                                 <a href="#" class="rounded-md bg-gray-900 px-3 py-2 text-sm font-medium text-white" aria-current="page">Browse Gallery</a>
-                                <a href="#" class="rounded-md px-3 py-2 text-sm font-medium text-gray-300 hover:bg-gray-700 hover:text-white">Featured Artists</a>
+                                <a href="@Url.Action("AllArtist", "Guest")" class="rounded-md px-3 py-2 text-sm font-medium text-gray-300 hover:bg-gray-700 hover:text-white">Featured Artists</a>
                                 <a href="#" class="rounded-md px-3 py-2 text-sm font-medium text-gray-300 hover:bg-gray-700 hover:text-white">Categories</a>
                                 <a href="#" class="rounded-md px-3 py-2 text-sm font-medium text-gray-300 hover:bg-gray-700 hover:text-white">About</a>
                             </div>
